+++ changelog.md
@@ -1,3 +1,6 @@
+* Add FIO - Foundation for Interwallet Operability
+* Add Argoneum
+
 # 0.4.2
 
 * Fix ethereum private key format